implementare gestionare clienti

+++ PROIECTPOOMALL2/hipermarket.cpp
@@ -448,9 +448,9 @@ int Hipermarket::GetStoc(int id)
 	std::cout << "Produsul nu exista!" << std::endl;
 	return -1;
 }
-int Hipermarket::CalculValoareProduse() const
+double Hipermarket::CalculValoareProduse() const
 {
-	int suma = 0;
+	double suma = 0;
 	for (auto& g : m_catalog_mancare)
 	{
 		suma = suma + g.GetPret() * g.GetStoc();
