+++ PROIECTPOOMALL2/hipermarket.h
@@ -14,5 +14,5 @@ public:
 	virtual void AfiseazaAngajat() override;
 	virtual void ModificaStoc(int id, int cantitate) override;
 	virtual int GetStoc(int id) override;
-	virtual int CalculValoareProduse() const override;
+	virtual double CalculValoareProduse() const override;
 };
