+++ PROIECTPOOMALL2/magazin_electronice.cpp
@@ -86,9 +86,9 @@ int MagazinElectronice::GetStoc(int id)
 		std::cout << "Produsul nu exista!" << std::endl;
 		return -1;
 }
-int MagazinElectronice::CalculValoareProduse() const
+double MagazinElectronice::CalculValoareProduse() const
 {
-	int suma = 0;
+	double suma = 0;
 	for (auto& g : m_catalog_gadget)
 	{
 		suma = suma + g.GetPret() * g.GetStoc();
